[refactor] Swagger 보안 스키마를 Bean으로 관리 및 Authorize 순서 정렬
[fix] 자기 호출 트랜잭션 문제 해결을 위해 MatchSyncWorker 생성 및 동기화 메서드 분리
