+++ src/main/java/com/test/basic/lol/domain/team/TeamController.java
@@ -54,8 +54,8 @@ public class TeamController {
 
     @PostMapping("/sync")
     @PreAuthorize("hasAuthority('ADMIN')")
-    @SecurityRequirement(name = "IgnoreCSRF")
-    @SecurityRequirement(name = "CSRF")
+    @SecurityRequirement(name = "03_CSRF")
+    @SecurityRequirement(name = "04_IgnoreCSRF")
     @Operation(summary = "LOL 팀 정보 수동 동기화", description = "LOL 팀 정보 수동 동기화 API")
     public ResponseEntity<String> syncTeams() {
         logger.info("==================== 팀 정보 수동 동기화 작업 시작 ====================");
